Enable caching and buffering for all of physical RAM in the decompressor.
From: Mike Kasick <mike@kasick.org>
Date: Tue, 18 Oct 2011 21:29:44 -0400

Enable caching and buffering for all of physical RAM in the decompressor.

Old method is to enable caching and buffering only for the 256 MB at the
start of the decompressor image.  This makes a hardboot-kexec'd kernel very
slow to decompress since the decompressor is located far above the kernel
destination.  This patch reduces boot time from 35 to 8 seconds.